\name{calcDist}
\alias{calcDist}
\title{calcDist a lower-level helper function}
\usage{
calcDist(x1, y1, x2, y2)
}
\arguments{
  \item{x1}{x coordinates of points}

  \item{y1}{y coordinates of points}

  \item{x2}{reference point x coordinate}

  \item{y2}{reference point y coordinate}
}
\description{
Wrapper function which calls some underlying C code.
}

