\name{ConCode}
\alias{ConCode}
\title{View Condition Codes}
\usage{
ConCode(site = "BCI")

ConCode(site = "BCI")
}
\arguments{
  \item{site}{Which site? Can be either "BCI", "PASOH",
  ...}

  \item{site}{Which site? Can be either "BCI", "PASOH",
  ...}
}
\description{
\code{ConCode} - Gives an overview of the condition codes
for FDP plot data (only BCI currently).

\code{ConCode} - Gives an overview of the condition codes
for FDP plot data (only BCI currently).
}
\details{
Add condition codes as a help page below
}
\author{
Marco D. Visser

Marco D. Visser
}

