\name{CrowdStat.count}
\alias{CrowdStat.count}
\title{CrowdStat.count}
\usage{
  CrowdStat.count(distancemat, r)
}
\arguments{
  \item{distancemat}{a matrix containing the distances
  between all individuals}

  \item{r}{a vector containing all distances at which the
  below statistic must be calculated}
}
\description{
  primitive function which counts individual surrounding a
  given individual in the census data
}
\author{
  Marco D. Visser not exported
}

