\name{CrowdStat.nddcount}
\alias{CrowdStat.nddcount}
\title{CrowdStat.nddcount}
\usage{
  CrowdStat.nddcount(FdpData, r)
}
\arguments{
  \item{distancemat}{a matrix containing the distances
  between all individuals}

  \item{r}{a vector containing all distances at which the
  below statistic must be calculated}
}
\description{
  primitive function which counts points (trees)
  surrounding a given focal point (tree) in the census data
}
\author{
  Marco D. Visser
}

