\name{EMexp}
\alias{EMdexp}
\alias{EMexp}
\alias{EMiexp}
\title{the EMexp set
\code{EMexp} - A family of exponential functions,
predicting size at time t, from initial size M0, and a set of
parameters. The set includes an inverse and differential, which
predict age at size M, and growth rates dM/dtime for a size M.}
\usage{
EMexp(M0, r, time)

EMiexp(M0, r, time)

EMdexp(M, r)
}
\arguments{
  \item{time}{time since initial size}

  \item{M}{mass}

  \item{M0}{initial mass}

  \item{r}{growth rate}
}
\description{
the EMexp set \code{EMexp} - A family of exponential
functions, predicting size at time t, from initial size M0,
and a set of parameters. The set includes an inverse and
differential, which predict age at size M, and growth rates
dM/dtime for a size M.
}
\author{
Marco D. Visser

Marco D. Visser

Marco D. Visser
}

