\name{EMilogis3k}
\alias{EMilogis3k}
\title{3 parameter logistic with initial size parameter}
\usage{
EMilogis3k(M, M0, K, r)
}
\arguments{
  \item{time}{time since initial size}

  \item{M}{mass}

  \item{M0}{initial mass}

  \item{K}{asymptotic size}

  \item{r}{growth rate}
}
\description{
\code{EElogis} - A family of 3 parameter logistic function
}
\author{
Marco D. Visser
}

