% Generated by roxygen2 (4.0.1): do not edit by hand
\name{FitStats}
\alias{FitStats}
\title{Calculate fit statistics}
\usage{
FitStats(fitvalues, stat = "delta")
}
\arguments{
\item{fitvalues}{a list of AIC, DIC or BIC values}

\item{stat}{do you wish to calculate 'delta' values, 'weights' or ...
list to grow.
when the model is returned?}

\item{...}{additional parameters to be passed to \code{jags.model}}
}
\description{
\code{FitStats} - Calculates various fit statistics, including
Delta AIC/DIC/BIC values from a set of stats, AIC weighst and ...
list to grow.
}
\author{
Marco D. Visser
}

