% Generated by roxygen2 (4.0.1): do not edit by hand
\name{GROpow}
\alias{GROdpow}
\alias{GROipow}
\alias{GROpow}
\title{the GROpow set
\code{GROpow} - A family of power functions for size
growth modelling, predicting size at time t, from initial size M0,
and a set of
parameters. The set includes an inverse and differential, which
predict age at size M, and growth rates dM/dtime for a size M.}
\usage{
GROpow(M0, beta, r, time, betas = NULL)

GROipow(M, M0, r, beta, time, betas = NULL)

GROdpow(M, r, beta, betas = NULL)
}
\arguments{
\item{time}{time since initial size}

\item{M}{mass}

\item{M0}{initial mass}

\item{r}{growth rate}

\item{beta}{scaling factor}
}
\description{
the GROpow set
\code{GROpow} - A family of power functions for size
growth modelling, predicting size at time t, from initial size M0,
and a set of
parameters. The set includes an inverse and differential, which
predict age at size M, and growth rates dM/dtime for a size M.
}
\author{
Marco D. Visser

Marco D. Visser

Marco D. Visser
}

