% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/JAGSmodels.R
\name{loadModel}
\alias{loadModel}
\title{Set of common Jags models

\code{loadModel} - Contains a set of jags models
which include:
 - univariate probability density function (weibull,
 lognormal, normal, exponential).
- more to come}
\usage{
loadModel(model = "weibull", class = "pdf")
}
\arguments{
\item{model}{the selected model}

\item{class}{the model class (pdf,glm,lm)}
}
\description{
load a standard Jags models
}
\details{
\code{loadModel} - load a predifined BUGS model
}
\author{
Marco D. Visser
}

