\name{calcDistMat}
\alias{calcDistMat}
\title{calcDistMat}
\usage{
  calcDistMat(pointdata, xy = c("x", "y"))
}
\arguments{
  \item{pointdata}{a dataset containing x and y
  coordinates}

  \item{xy}{column names containing coordinates}
}
\description{
  Generic function to calculate distances between all point
  is a datatset containing x and y coordinates
}
\author{
  Marco D. Visser
}

