% Generated by roxygen2 (4.0.0): do not edit by hand
\name{calculateCrowding}
\alias{calculateCrowding}
\title{Calculate crowding statistics}
\usage{
calculateCrowding(FdpData = list(bci.full6, bci.full7), refdata = NULL,
  r = seq(1, 50, 10), statistic = "count", ...)
}
\arguments{
\item{FdpData}{A list of fdp census datasets on which the
calculation must be conducted}

\item{refdata}{a second dataset including points for
which crowding statistics based on the FdpData must
be calculated. This option is ignored if NULL.}

\item{r}{a vector containing all distances at which the
below statistic must be calculated. Used if relevant.}

\item{statsistic}{character defining which method to use.
Methods include: 1) "count", simple count of neighbours
within r 2) "distance index", a distance weighted function
3) "gravitation", a size and distance weighted function}

\item{...}{further arguments to be passed to statistic
calculation}
}
\description{
\code{calculateCrowding} - Calculates crowding statistics
for each individual alive, based on the surrounding
environment, in each included census for the BCI
plot data. Alternatively, if a second data set is included
(refdata) it calculates crowding statistics for all points
in the second dataset based on all points in each FDP census
included.
}
\author{
Marco D. Visser
}

