% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NddMeassuresFunctions.R
\name{CrowdStat.ndd}
\alias{CrowdStat.ndd}
\title{CrowdStat.nddcount}
\usage{
\method{CrowdStat}{ndd}(FdpData, r)
}
\arguments{
\item{r}{a vector containing all distances at which the
below statistic must be calculated}

\item{distancemat}{a matrix containing the distances
between all individuals}
}
\description{
primitive function which counts points (trees)
surrounding a given focal point (tree) in the census data
}
\author{
Marco D. Visser
}

