\name{GROexp}
\alias{GROdexp}
\alias{GROexp}
\alias{GROiexp}
\title{the GROexp set
\code{GROexp} - A family of exponential functions for size
growth modelling. These models predict size at time t,
from initial size M0, and a set of
parameters. The set includes an inverse and differential, which
predict age at size M, and growth rates dM/dtime for a size M.}
\usage{
GROexp(M0, r, time)

GROiexp(M0, r, time)

GROdexp(M, r)
}
\arguments{
  \item{time}{time since initial size}

  \item{M}{mass}

  \item{M0}{initial mass}

  \item{r}{growth rate}
}
\description{
the GROexp set \code{GROexp} - A family of exponential
functions for size growth modelling. These models predict
size at time t, from initial size M0, and a set of
parameters. The set includes an inverse and differential,
which predict age at size M, and growth rates dM/dtime for
a size M.
}
\author{
Marco D. Visser

Marco D. Visser

Marco D. Visser
}

