% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/EcologyModels.R
\name{GROlinear}
\alias{GROdlinear}
\alias{GROilinear}
\alias{GROlinear}
\title{the GROlinear set

\code{GROlinear} - A family of linear functions for size
growth modelling. These models predict size at time t,
from initial size M0, and a set of
parameters. The set includes an inverse and differential, which
predict age at size M, and growth rates dM/dtime for a size M.}
\usage{
GROlinear(M0, r, time = 1, betas = NULL)

GROilinear(M, M0, r, betas = NULL)

GROdlinear(M, r, betas = NULL)
}
\arguments{
\item{M0}{initial mass}

\item{r}{growth rate}

\item{time}{time since initial size}

\item{M}{mass}
}
\description{
the GROlinear set

\code{GROlinear} - A family of linear functions for size
growth modelling. These models predict size at time t,
from initial size M0, and a set of
parameters. The set includes an inverse and differential, which
predict age at size M, and growth rates dM/dtime for a size M.
}
\author{
Marco D. Visser

Marco D. Visser

Marco D. Visser
}

