\name{summary.FDPmodel}
\alias{summary.FDPmodel}
\title{Summarize models and fitstatistics}
\usage{
summary.FDPmodel(modelfit, ...)
}
\arguments{
  \item{modelfit}{A list of FDPmodel objects returned by
  FitConv (e.g. FitConv run on multiple species)}

  \item{...}{additional parameters}
}
\description{
\code{summary.FDPmodel} - summarizes of a list of
previously fit models returned by \code{FitConv}
}
\author{
Marco D. Visser
}

